mythfilldatabase: rename commandlineparser.* to mythfilldatabase_…
