+++ mythtv/programs/mythfilldatabase/mythfilldatabase.pro
@@ -19,7 +19,7 @@ DEPENDPATH  += ../../libs/libmythtv/mpeg
 # Input
 HEADERS += filldata.h   channeldata.h
 HEADERS += xmltvparser.h
-HEADERS += fillutil.h   commandlineparser.h
+HEADERS += fillutil.h   mythfilldatabase_commandlineparser.h
 SOURCES += filldata.cpp channeldata.cpp
 SOURCES += xmltvparser.cpp fillutil.cpp
-SOURCES += mythfilldatabase.cpp     commandlineparser.cpp
+SOURCES += mythfilldatabase.cpp     mythfilldatabase_commandlineparser.cpp
